Radar

+++ body/classes/system/GameField.h
@@ -105,7 +105,7 @@ Array<int> deeps;
     Cell *drawCell = nullptr;
     string path;
     int type;
-    float miniMapUpdateVer = 0;
+    int miniMapUpdateVer = 0;
     float gabarit = 0;
     float screenWidth = 720;
     float screenHeight = 720;
@@ -143,8 +143,8 @@ Array<int> deeps;
     bool miniMapClick = false;
     float miniMapWinGab = 0;
     float mcs = 0.0f;
-    int dx = 0;
-    int dy = 0;
+    double dx = 0;
+    double dy = 0;
     //////////////////////////// <= miniMap
 
     ///////////////// => fieldClick
